third_party/blink/web_tests/external/wpt/css/cssom/idlharness-expected.txt

This is a testharness.js-based test.
Found 31 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] CSSStyleRule interface: existence and properties of interface object
  assert_equals: prototype of CSSStyleRule is not CSSGroupingRule expected function "function CSSGroupingRule() { [native code] }" but got function "function CSSRule() { [native code] }"
[FAIL] CSSStyleRule interface: existence and properties of interface prototype object
  assert_equals: prototype of CSSStyleRule.prototype is not CSSGroupingRule.prototype expected object "[object CSSGroupingRule]" but got object "[object CSSRule]"
[FAIL] CSSImportRule interface: sheet.cssRules[0] must inherit property "layerName" with the proper type
  assert_equals: expected "string" but got "object"
[FAIL] CSSImportRule interface: sheet.cssRules[0] must inherit property "supportsText" with the proper type
  assert_equals: expected "string" but got "object"
[FAIL] CSSPageDescriptors interface: existence and properties of interface object
  assert_own_property: self does not have own property "CSSPageDescriptors" expected property "CSSPageDescriptors" missing
[FAIL] CSSPageDescriptors interface object length
  assert_own_property: self does not have own property "CSSPageDescriptors" expected property "CSSPageDescriptors" missing
[FAIL] CSSPageDescriptors interface object name
  assert_own_property: self does not have own property "CSSPageDescriptors" expected property "CSSPageDescriptors" missing
[FAIL] CSSPageDescriptors interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "CSSPageDescriptors" expected property "CSSPageDescriptors" missing
[FAIL] CSSPageDescriptors interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "CSSPageDescriptors" expected property "CSSPageDescriptors" missing
[FAIL] CSSPageDescriptors interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "CSSPageDescriptors" expected property "CSSPageDescriptors" missing
[FAIL] CSSPageDescriptors interface: attribute margin
  assert_own_property: self does not have own property "CSSPageDescriptors" expected property "CSSPageDescriptors" missing
[FAIL] CSSPageDescriptors interface: attribute marginTop
  assert_own_property: self does not have own property "CSSPageDescriptors" expected property "CSSPageDescriptors" missing
[FAIL] CSSPageDescriptors interface: attribute marginRight
  assert_own_property: self does not have own property "CSSPageDescriptors" expected property "CSSPageDescriptors" missing
[FAIL] CSSPageDescriptors interface: attribute marginBottom
  assert_own_property: self does not have own property "CSSPageDescriptors" expected property "CSSPageDescriptors" missing
[FAIL] CSSPageDescriptors interface: attribute marginLeft
  assert_own_property: self does not have own property "CSSPageDescriptors" expected property "CSSPageDescriptors" missing
[FAIL] CSSPageDescriptors interface: attribute margin-top
  assert_own_property: self does not have own property "CSSPageDescriptors" expected property "CSSPageDescriptors" missing
[FAIL] CSSPageDescriptors interface: attribute margin-right
  assert_own_property: self does not have own property "CSSPageDescriptors" expected property "CSSPageDescriptors" missing
[FAIL] CSSPageDescriptors interface: attribute margin-bottom
  assert_own_property: self does not have own property "CSSPageDescriptors" expected property "CSSPageDescriptors" missing
[FAIL] CSSPageDescriptors interface: attribute margin-left
  assert_own_property: self does not have own property "CSSPageDescriptors" expected property "CSSPageDescriptors" missing
[FAIL] CSSPageDescriptors interface: attribute size
  assert_own_property: self does not have own property "CSSPageDescriptors" expected property "CSSPageDescriptors" missing
[FAIL] CSSPageDescriptors interface: attribute pageOrientation
  assert_own_property: self does not have own property "CSSPageDescriptors" expected property "CSSPageDescriptors" missing
[FAIL] CSSPageDescriptors interface: attribute page-orientation
  assert_own_property: self does not have own property "CSSPageDescriptors" expected property "CSSPageDescriptors" missing
[FAIL] CSSPageDescriptors interface: attribute marks
  assert_own_property: self does not have own property "CSSPageDescriptors" expected property "CSSPageDescriptors" missing
[FAIL] CSSPageDescriptors interface: attribute bleed
  assert_own_property: self does not have own property "CSSPageDescriptors" expected property "CSSPageDescriptors" missing
[FAIL] CSSStyleProperties interface: existence and properties of interface object
  assert_own_property: self does not have own property "CSSStyleProperties" expected property "CSSStyleProperties" missing
[FAIL] CSSStyleProperties interface object length
  assert_own_property: self does not have own property "CSSStyleProperties" expected property "CSSStyleProperties" missing
[FAIL] CSSStyleProperties interface object name
  assert_own_property: self does not have own property "CSSStyleProperties" expected property "CSSStyleProperties" missing
[FAIL] CSSStyleProperties interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "CSSStyleProperties" expected property "CSSStyleProperties" missing
[FAIL] CSSStyleProperties interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "CSSStyleProperties" expected property "CSSStyleProperties" missing
[FAIL] CSSStyleProperties interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "CSSStyleProperties" expected property "CSSStyleProperties" missing
[FAIL] CSSStyleProperties interface: attribute cssFloat
  assert_own_property: self does not have own property "CSSStyleProperties" expected property "CSSStyleProperties" missing
Harness: the test ran to completion.
